% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpa.R
\name{LPA_deter}
\alias{LPA_deter}
\title{Deterministic Larvae-Pupae-Adult model}
\usage{
LPA_deter(
  N0 = c(larvae = 75, pupae = 35, adults = 60),
  b,
  c_el,
  c_ea,
  c_pa,
  mu_l,
  mu_a,
  V
)
}
\arguments{
\item{N0}{vector of initial population size}

\item{b}{eggs laid per adult}

\item{c_el}{rate of cannabilism of larvae on eggs}

\item{c_ea}{rate of cannabilism of adults on eggs}

\item{c_pa}{rate of cannabilism of adults on pupae}

\item{mu_l}{probability of larval death in 2 weeks}

\item{mu_a}{probability of adult death in 2 weeks}

\item{V}{volume of habitat (1 = 20 g)}
}
\value{
vector of stage numbers
}
\description{
Deterministic Larvae-Pupae-Adult model
}
\examples{
# Parameters of control replicate A from Desharnais and Costantino (1980)
LPA_deter(b = 11.67, c_el = 0.0093, c_ea = 0.0110, c_pa = 0.0178, mu_l = 0.5129, mu_a = 0.1108, V = 1)
iterate(parms = controlA, N0 = c(larvae = 75, pupae = 35, adults = 60), popfun = LPA_deter)
}
