% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_beetles.R
\name{sample_beetles}
\alias{sample_beetles}
\title{Title}
\usage{
sample_beetles(N, V, n, v, replacement = TRUE)
}
\arguments{
\item{N}{vector of larvae, pupae, adult to draw sample from}

\item{V}{total volume of habitat in units of 20 g}

\item{n}{vector of number of samples to take from each stage}

\item{v}{volume of sample}

\item{replacement}{(logical) sample with or without replacement}
}
\value{
data frame with a column for each stage sampled, and a row for each sample.
}
\description{
Title
}
\examples{
sample_beetles(c(larvae = 75, pupae = 35, adults = 150), V = 5, n = 5, v = 0.15)
beetles <- iterate(parms = controlA, N0 = c(larvae = 75, pupae = 35, adults = 60), popfun = LPA_deter)
## Not run:
samples <- purrr::map_dfr(1:20, function(t)sample_beetles(unlist(beetles[t, 9:11]), V = 1, n = c(0,0,5), v = 0.15), .id = "t")

## End(Not run)
}
